+++ src/findfonts.jl
@@ -1,72 +1,3 @@
-mutable struct FTFont
-    ft_ptr::FreeType.FT_FaceRec
-    pixel_size::Int
-    use_cache::Bool
-    cache::Dict{Char, FontExtent{Float32}}
-    function FTFont(ft_ptr::FreeType.FT_FaceRec, pixel_size::Int=64, use_cache::Bool=true)
-        cache = Dict{Char, FontExtent{Float32}}()
-        face = new(ft_ptr, pixel_size, use_cache, cache)
-        finalizer(FT_Done_Face, face)
-        FT_Set_Pixel_Sizes(face, pixel_size, 0);
-        return face
-    end
-end
-
-
-function FTFont(path::String)
-    face = Ref{FT_Face}(C_NULL)
-    err = FT_New_Face(ftlib[1], path, Int32(faceindex), face)
-    if err != 0
-        error("Couldn't load font $facename with error $err")
-    end
-    return FTFont(face[])
-end
-
-function FTFont(font_ptr::Ptr{FreeType.FT_FaceRec})
-    face_rect = unsafe_load(font_ptr)
-    return FTFont(face_rect)
-end
-
-# C interop
-function Base.cconvert(::Type{FreeType.FT_Face}, font::FTFont)
-    return font
-end
-
-function Base.unsafe_convert(::Type{FreeType.FT_Face}, font::FTFont)
-    ptr = Base.pointer_from_objref(font)
-    return convert(FreeType.FT_Face, ptr)
-end
-
-function Base.propertynames(font::FTFont)
-    return fieldnames(FreeType.FT_FaceRec)
-end
-
-function Base.getproperty(font::FTFont, fieldname::Symbol)
-    fontrect = getfield(font, :ft_ptr)
-    field = getfield(fontrect, fieldname)
-    if field isa Ptr{FT_String}
-        return unsafe_string(field)
-    # Some fields segfault with unsafe_load...Lets find out which another day :D
-    elseif field isa Ptr{FreeType.LibFreeType.FT_GlyphSlotRec}
-        return unsafe_load(field)
-    else
-        return field
-    end
-end
-
-get_pixelsizes(face::FTFont) = (get_pixelsize(face), get_pixelsize(face))
-get_pixelsize(face::FTFont) = getfield(face, :pixel_size)
-reset_pixelsize(face::FTFont) = FT_Set_Pixel_Sizes(face, get_pixelsize(face), 0)
-
-function check_error(err, error_msg)
-    if err != 0
-        error(error_msg * " with error: $(err)")
-    end
-end
-
-use_cache(face::FTFont) = getfield(face, :use_cache)
-get_cache(face::FTFont) = getfield(face, :cache)
-
 if Sys.isapple()
     function _font_paths()
         [
@@ -98,52 +29,19 @@ else
     end
 end
 
-
-freetype_extensions() = (".FON", ".OTC", ".FNT", ".BDF", ".PFR", ".OTF", ".TTF", ".TTC", ".CFF", ".WOFF")
-
-function freetype_can_read(font::String)
-    fontname, ext = splitext(font)
-    uppercase(ext) in freetype_extensions()
-end
-
-function loaded_faces()
-    if isempty(loaded_fonts)
-        for path in fontpaths()
-            for font in readdir(path)
-                # There doesn't really seem to be a reliable pattern here.
-                # there are fonts that should be supported and dont load
-                # and fonts with an extension not on the FreeType website, which
-                # load just fine. So we just try catch it!
-                #freetype_can_read(font) || continue
-                fpath = joinpath(path, font)
-                try
-                    push!(loaded_fonts, newface(fpath)[1])
-                catch
-                end
-            end
-        end
-    end
-    return loaded_fonts
-end
-
 family_name(x::String) = replace(lowercase(x), ' ' => "") # normalize
 
-function family_name(x)
-    fname = x.family_name
-    fname == C_NULL && return ""
-    family_name(unsafe_string(fname))
+function family_name(x::FTFont)
+    family_name(x.family_name)
 end
 
-function style_name(x)
-    sname = x.style_name
-    sname == C_NULL && return ""
-    lowercase(unsafe_string(sname))
+function style_name(x::FTFont)
+    lowercase(x.style_name)
 end
 
-function match_font(face, name, italic, bold)
-    ft_rect = unsafe_load(face)
-    fname = family_name(ft_rect)
-    sname = style_name(ft_rect)
+function match_font(face::FTFont, name, italic, bold)
+    fname = family_name(face)
+    sname = style_name(face)
     italic = italic == (sname == "italic")
     bold = bold == (sname == "bold")
     perfect_match = (fname == name) && italic && bold
@@ -154,7 +52,7 @@ end
 
 function try_load(fpath)
     try
-        newface(fpath)[]
+        return FTFont(fpath)
     catch e
         return nothing
     end
@@ -162,12 +60,12 @@ end
 
 function findfont(
         name::String;
-        italic = false, bold = false, additional_fonts::String = ""
+        italic::Bool=false, bold::Bool=false, additional_fonts::String=""
     )
     font_folders = copy(fontpaths())
     normalized_name = family_name(name)
     isempty(additional_fonts) || pushfirst!(font_folders, additional_fonts)
-    candidates = Pair{Ptr{FreeType.FT_FaceRec}, Int}[]
+    candidates = Pair{FTFont, Int}[]
     for folder in font_folders
         for font in readdir(folder)
             fpath = joinpath(folder, font)
@@ -179,16 +77,13 @@ function findfont(
             perfect_match && return face
             if fuzzy_match
                 push!(candidates, face => score)
-            else
-                FT_Done_Face(face)
             end
         end
     end
     if !isempty(candidates)
-        sort!(candidates, by = last)
+        sort!(candidates; by=last)
         final_candidate = pop!(candidates)
-        foreach(x-> FT_Done_Face(x[1]), candidates)
-        return FTFont(final_candidate[1])
+        return final_candidate[1]
     end
     return nothing
 end
